WIP(web-diary): bpm edited content
affects: @practice/web-diary

ISSUES CLOSED: #81
