--- package/list/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@gochicus/list",
-  "version": "0.0.0",
-  "description": "item iteration",
-  "main": "index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\""
-  },
-  "author": "gochicus",
-  "license": "BSD-3-Clause",
-  "dependencies": {
-    "@atlantis-lab/tsconfig": "^0.1.12",
-    "@babel/core": "^7.14.6",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@types/react": "^17.0.11",
-    "react": "^17.0.2",
-    "styled-system": "^5.1.5",
-    "styled-tools": "^1.7.2",
-    "typescript": "^4.3.4"
-  },
-  "devDependencies": {
-    "@types/styled-system": "^5.1.11"
-  }
-}
