--- package/list/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@atlantis-lab/tsconfig/tsconfig.base.json",
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ES2015"
-  },
-  "lib": ["es2019.intl"],
-  "include": ["./"]
-}
