+++ packages/list/src/icons/src/StarIcon.tsx
