+++ packages/list/src/icons/src/TrashIcon.tsx
