--- practice/web-diary/ui/list/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@ui/list",
-  "version": "0.0.0",
-  "license": "BSD 3-Clause",
-  "main": "src/index.tsx",
-  "dependencies": {
-    "@store/post-data": "0.0.0",
-    "@ui/background": "0.0.0",
-    "@ui/button": "0.0.0",
-    "@ui/condition": "0.0.0",
-    "@ui/icons": "0.0.0",
-    "@ui/layout": "0.0.0",
-    "@ui/text": "0.0.0",
-    "@ui/theme": "0.0.0",
-    "react": "^17.0.1"
-  }
-}
