--- practice/web-diary/ui/list/src/index.tsx
@@ -1,2 +1,0 @@
-export * from './List'
-export * from './Item'
